Include arguments with the artifacts (#2260)
# Motivation
Deployments of the nns-dapp now need arguments. It is good if a
reference set of arguments is created in CI to check against, in the
same way that local builds of the wasm are checked against CI builds.

# Changes
* Include the build configuration in the CI artifacts.  In particular:
* `deployment-config.json` - the config from which all other config is
derived.
* `nns-dapp-arg.did` - the text format of the argument, used when
deploying directly
* `nns-dapp-arg.bin` - the binary form of the argument, used when
deploying by proposal
* `frontend-config.sh` - the contents of the .env file, used when doing
front end develpment.

# Tests
* See the artifacts when CI runs. As an added bonus, the wasm may no
longer be double-unzipped by mac. Let's see.

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>