GIX-1440: NNS Dapp uses inserted env vars if possible (#2230)
# Motivation

In order for the NNS Dapp wasm to be agnostic, we introduced arguments
when installing the wasm that create an element in the html with all the
environment configuration.

This PR adds the functionality in the frontend assets to use that
configuration first.

I deployed to
https://3cswx-rqaaa-aaaaa-aabyq-cai.nnsdapp.testnet.dfinity.network/accounts/?u=3cswx-rqaaa-aaaaa-aabyq-cai
a wasm created for another tesnet. I get the calls to the proper
canister id, set on installation.

Yet, this is not working because the CSP are not set.

# Changes

* New environment util `getEnvVars` that returns an object with the
configuration. This util either reads the env vars from build or html
depending on whether is local development or not.
* Use new util in all the environment variables.

# Tests

* Only build changes.