Implement waitForAbsent on JestPageObjectElement (#2288)
# Motivation

We use page objects for both e2e tests and jest unit tests.
The e2e tests and unit tests that use page objects will overlap more and
more and it is good to use the same patterns in both of them.

Also, JestPageObjectElement.isPresent wasn't using the most recent state
of the DOM to determine if an element is present.

# Changes

1. Add a private method `getRootAndFullSelector` to
`JestPageObjectElement` to get the furthest available ancestor of the
current element and the full selector to point from it to the current
element.
2. Use `getRootAndFullSelector` to implement `isPresent` and update
this.element.
3. Use `isPresent` to implement both `waitFor` and `waitForAbsent`.

`waitForAbsent` currently isn't used by any unit tests, but the new
implementation is so straightforward that it will likely work and some
testing from another branch suggests that it does.

# Tests

`npm run test`