GIX-1481: Fetch the latest reward event (#2292)
# Motivation

We want to show the last time that maturity was distrubuted.

This PR: It implements the api functions to get latest reward events and
sets up the service.

Upcoming PRs: Store the event in a store and use it in the component.

# Changes

* Upgrade next dependencies.
* New governance api function `queryLastestRewardEvent`.
* Add new method in Governance Api Service `queryLastestRewardEvent`.
For now, no caching is done.
* Setup reward service `loadLatestRewardEvent` and use it in
NnsNeuronDetail page mount.

# Tests

* Test new api function.
* Test new governance api service method.
* Mock api function where needed to pass current tests.