Remove mock conversion rate feature flag (#2314)
# Motivation
The backend has a feature flag that now does nothing.
`mock_conversion_rate` used to be used as follows, but now there is no
code whatsoever that depends on it. However it is problematic because
the build encodes the feature flags, so we have different backend builds
for mainnet and testnets.
```
+    #[cfg(feature = "mock_conversion_rate")]
+    {
+        100_000_000_000_000 // Assume a conversion rate of 100T cycles per 1 ICP.
+    }
+
+    #[cfg(not(feature = "mock_conversion_rate"))]
+    {
+        const CYCLES_PER_XDR: u64 = 1_000_000_000_000;
```

# Changes

<!-- List the changes that have been developed -->

# Tests
See CI

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>