test: Get aggregator args & compare with deployment (#2309)
# Motivation
There are no tests for sns canister configuration.

# Changes
- Add a query method to get the canister config.
- Add a test to verify that the config is as specified.
- The test passes for configuration set with `reconfigure(..)` and fails
for configuration set with `dfx canister install()`. Amusingly,
reconfigure() is a custom method, seemed risky, and was exercised in
tests extensively. It is the safe standard method causing trouble.
- Examination shows that the arguments are not passed to the canister at
all, so the issue is somehow related to the way dfx tries to send the
arguments.

# TODO
- Find out why dfx is not passing the argument & fix.
- Check that when canister arguments are received, they also take
effect. E.g. that if a certain update frequency is specified in the
config, the `sns_aggregator` really does get data with that frequency.

# Tests
- A test is included in this PR.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>