Nightly tag (#2346)
# Motivation
We wish to test a recent build (not prod) against the rest of the IC
test environment.

## Requirements
* Nightly release
* No need to make a new release if there is an existing release
* The nns-dapp wasm should be available with the commit in the filename.

# Changes
* To the nightly workflow, add a job that tags the head of main with
`nightly-YYY-mm-dd`
* In the docker-build job that creates releases:
* Make a copy of the wasm that includes the commit in the name and
include it in the release artefacts.
  * Make nightly releases public, so that the wasms are accessible

# Tests
* I have triggered the workflow manually by pushing to the develop
branch. It is hard to get workflows to get tags, so I have not been able
to confirm that no new releases are created if there are no changes. I
have been able to confirm that the nightly job creates tags and that the
tags trigger workflows.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>
Co-authored-by: Nightly GitHub Action <nobody@dfinity.org>