Chore: Use parent data-tid in KeyValuePairInfo (#2355)
# Motivation

The page object of the KeyValuePairInfo was relying in the parent to get
the description because the data-tid was not added at the root.

The PR in GIX Components:
https://github.com/dfinity/gix-components/pull/199 sets the test id of
KeyValuePairInfo in the root component.

# Changes

* Upgrade gix-components library.
* Change the way KeyValuePairInfoPo accesses the description text. There
is no need to keep a reference to the parent.

# Tests

Changes only in test implementations, not no test changes.