Chore: Add TVL canister id to config (#2354)
# Motivation

NNS Dapp testnet tries to get data from mainnet. Because of that, errors
are thrown in the console on testnet.

With errors in the console is hard to differentiate real errors and ok
errors.

The data is from the TVL canister. We also might want to deploy a TVL
canister to a testnet and check that all works as expected.

Therefore, this PR adds the TVL canister to the configuration.

# Changes

* Add TVL_CANISTER_ID in `config.sh`.
* Add `tvlCanisterId` key to envVars exposed in frontend
* Add parameter of canisterId to api `queryTVL` and use the `HOST` env
var instead of mainnet host.
* Use env var `tvlCanisterId` in environment variable `TVL_CANISTER_ID`.
* Check that TVL_CANISTER_ID is present in service `queryTVL`.
* Convert empty string ot undefined also the env vars from HTML. Use a
new helper `mapEmtpyStringsToUndefined` for that.
* Add `tvlCanisterId` to envVars from HTML and build step.

# Tests

* Add TVL_CANISTER_ID to test for `config.sh`
* Fix broken tests after changing tvl api and service
* Add `tvlCanisterId` key to jest-setup