Include network in arg file paths (#2362)
# Motivation
How is it possible to mess up when deploying to production? One answer:
It is possible to deploy testnet artefacts to mainnet. Testnet and
mainnet now use the same wasm, however it is possible to accidentally
use testnet args on mainnet, especially when we test deployments
beforehand on local and static testnets.

# Changes
- Rename nns-dapp-args.did and nns-dapp-args.bin to
`nns-dapp-args-${$DFX_NETWORK}.did` and `...bin`

# Tests
- [x]  CI
- [x] Manually check that mainnet releases contain the right network and
use the right arg files - see log in comment below.
- [x] Similarly check releases by proposal to a testnet - see log in
comment below.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>