Chore: Avoid source duplication env vars testing (#2374)
# Motivation

Jest setup was mocking twice some variables. For example, DFX_NETWORK
was mocked in the env-vars as well as in environment.constants.

This PR splits environment.constants.ts in two: the flags that depend on
the env vars, and the ones that not.

# Changes

Main changes:
* New file enviroment-flags.constants.ts
* Remove constants from environment.constants.ts

Minor changes:
* Change imports that used vars from environment.constants.ts and moved
to enviroment-flags.constants.ts

# Tests

* Mock enviroment-flags.constants.ts in jest-setup.ts