GIX-1416: Set icrc1_total_supply as required (#2383)
# Motivation

A few weeks ago we added a new field in the SNS aggregator. The NNS Dapp
upgraded first and to handle backwards compatibility we set the field as
optional.

Now that SNS aggregator was upgraded we can set the field
`icrc1_total_supply` as required.

# Changes

Main change:
* Set field icrc1_total_supply required in the data from aggregator
`CachedSnsMetadataDto`.

Minor change:
* No need to filter out by icrc1_total_supply not defined when setting
the `snsTotalTokenSupplyStore` from SNS aggregator.

# Tests

* Add the new field in the SNS Aggregator mocks.
* Remove icrc1_total_supply specific test cases. The field is now
required and expected with the mock.