E2E: Disable Firefox until I can debug. (#2397)
# Motivation

E2E tests are currently quite flaky because of a combination of being
slow and race conditions because of queryAndUpdate. The Firefox test
seem to be especially affected by this.
Let's disable them until I have time to look into the race conditions
and tweak the timeouts.

# Changes

Disable Firefox in Playwright config.

# Tests

None