Update HOWTO.md instruction to add a new NNS function proposal type (#2405)
# Motivation

I just added a new NNS function in
https://github.com/dfinity/nns-dapp/pull/2403 and I figured out a number
of things and wanted to document them here.
I also struggled a bit with the existing instructions so I tried to make
them clearer.
Please check that I didn't remove something useful just because I didn't
understand it.

# Changes

1. Overview of steps to add a new NNS function.
2. Detailed instruction of every step.
3. Use links that either fix a specific commit, or use a text search
fragment to point to a specific line. The existing links were broken
because the lines they pointed to changed.
4. Includes steps about testing the changes in addition to making the
changes

# Tests

I went through the steps to verify that they work for me.