Make ckBTC ledger API return a single account (#2433)
# Motivation

We will need to fetch the balance of another pseudo ckBTC account - the
withdrawal account - in PR #2430.
That's why this PR makes the ckBTC ledger API returns a single account
and extract the logic of fetching the main account into a new service.
That way we can in the other PR add a new function that will fetch the
account using the same API..

Refactor  only.