Display a warning if ckBTC to BTC conversion unexpectedly failed (#2418)
# Motivation

Display a warning on login if ckBTC to BTC conversion unexpectedly
failed - e.g. if user closes the browser after the transaction to the
ledger was made but the minter conversion was not triggered.

# PRs

- [x] #2420
- [x] https://github.com/dfinity/gix-components/pull/208

# Changes

- save block index in local storage after transfer to ledger in
conversion function
- remove block index in local storage after retrieve btc regardless if
success or error as long as ui is active
- on login, if local storage contains any block storage, display a
warning and clear local storage of the block indexes
- bump gix-cmp to display more lines in toast
- stringify bigint upon writing to local storage

# Screenshots

<img width="1535" alt="Capture d’écran 2023-05-02 à 15 39 02"
src="https://user-images.githubusercontent.com/16886711/235685578-67690f38-daee-4694-a8e1-c2d1940523f3.png">