Don't expose syncAccounts error handler (#2450)
# Motivation

`syncAccounts` has an errorHandler parameter but this is only used by
`initAccounts` in the same file.
By limiting the functionality we export out of the file it becomes
easier to refactor without breaking anything.

# Changes

1. Add `syncAccountsWithErrorHandler` (which is not exported) and called
by both `initAccounts` and `syncAccounts`.
2. Remove the optional errorHandler parameter from `syncAccounts`.
3. Remove the tests that test this parameter.

# Tests

Unused functionality is removed so the corresponding tests are removed.
`initAccounts` is still tested separately as it was before.