queuedStore to mitigate queryAndUpdate race condition (#2445)
# Motivation

There is a race condition with queryAndUpdate. If you call
queryAndUpdate again before the "update" from the first queryAndUpdate
responds, then when the first update does respond, it can overwrite
responses from the second queryAndUpdate with old data.
This is causing end-to-end tests to be flaky.

To mitigate this I implemented a "store wrapper" which keeps the pending
mutations in a queue. Then if an older mutation responds later, it is
placed in the correct place in the queue and all mutations are reapplied
in the correct order.
By "correct" order, I mean the order in which the original
queryAndUpdate calls happened.
Whether this is actually the correct order, we can't know unless we get
the block height in the response, which we currently don't.

To match responses from the same queryAndUpdate call, the caller is
required to request a one-time use store and apply to it just the
mutations corresponding to that queryAndUpdate.

# Changes

1. Implemented `queuedStore` which wraps a writable store and applies
the queue mechanism described above.
2. Changed accountsStore to use queuedStore.
3. Updated accounts service to use the new accountsStore API.
4. Temporarily added methods for the original API of accountsStore for
backwards compatibility until all existing tests have been updated.

# Tests

Added tests for queuedStore, accountsStore and accounts service.