Improve performance by not necessary getting the withdrawal account again (#2459)
# Motivation

If the withdrawal account has already been loaded in store in a
certified way, we can reuse this account to proceed with the ckBTC to
BTC conversion and skip an update call.

If not yet loaded, then we proceed as currently and we call the minter
to get the account.

# Notes

I decided to not overengineer this process - i.e. if there isn't an
account in store and if we fetch it, this process won't save for the
moment the resulting account in store. It isn't the goal of the function
and it doesn't arm the UI/UX to not do so.

# Changes

- in `convertCkBTCToBtc`, check for account, if present use it else
continue as existing