Load next transactions after refresh balance (#2483)
# Motivation

On ckBTC wallet page there is newly a "Refresh balance" button that used
to be in the modal.

When executed it reloads the account but, did not reload anymore the
transactions because well, it was moved without it.

This PR load the next transactions after such actions.

# Note

I just load the next transactions and do not reload everything because a
successful "refresh balance" produces a new transaction.

# Changes

- instead of `onMount` call load on any account changes