GIX-1555 Support conditions in AdditionalInfoForm (#2496)
# Motivation

We want to allow SNS params to include conditions that have to be agreed
to by swap participants.
This PR only updates the component that will display the conditions and
the checkbox.
It is not hooked up to anything yet.

# Changes

1. In `frontend/src/lib/components/sale/AdditionalInfoForm.svelte`
render conditions and a checkbox if conditions are passed as optional
props.
2. Add a page object for `AdditionalInfoForm`.
3. Add a `AdditionalInfoFormTest` component to assist in testing by
passing a svelte store to read out the value of the checkbox.
4. Add tests.

# Tests

1. Added a unit test.
2. Confirmed manually that nothing is displayed yet because the props
aren't passed to the component.