Chore: Refactor Participate Button with status helper (#2523)
# Motivation

The status of the Participate Button is becoming complex.

In this PR, we use the helper to get the participation button status in
the ParticipateButton component.

Not in this PR, manage the new scenarios with restricted countries.

# Changes

* ParticipateButton, use the button status from the helper instead of
other variables to render and not render.
* Change the type of the `userCountry` in util
`participateButtonStatus`.

# Tests

* Add new test case for "loading" in ParticipateButton spec file.
* Fix tests for participateButtonStatus after refactor.
* One test in ParticipateButton was broken and was passing for the wrong
reasons. I fixed it and added an extra expect to ensure the scenario is
the expected one.
* Fix one text in ProjectStatusSection. No commitment means loading
state, not present. Earlier, this case was not covered in the button.