Fix: npm run build (#2554)
# Motivation

`npm run build` is broken.

The problem: Code that depends on `window` was executed at build time.

Ideally: That component should not be executed at build time.

Workaround: Check whether the environment is the browser before calling
the problematic code.

# Changes

* Check whether we're in the browser in `GetTokens` before getting the
balance of the test account.

# Tests

This component is not tested because it's used only in testnets or loca
environments.